changed the path for trending and recommend test those files moved to components
